Refresh tasks in background
