+++ src/tasks.ts
@@ -1,9 +1,8 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
-import * as fs from 'fs';
 
 import { getIsAutoDetectionEnabled } from './config';
-import { GradleTasksClient, GradleTask, ServerMessage } from './server';
+import { GradleTasksClient, GradleTask, ServerMessage } from './client';
 
 export interface GradleTaskDefinition extends vscode.TaskDefinition {
   script: string;
@@ -18,7 +17,8 @@ export interface GradleTaskDefinition extends vscode.TaskDefinition {
 }
 
 let autoDetectOverride = false;
-let cachedTasks: Promise<vscode.Task[]> | undefined;
+let cachedTasks: vscode.Task[] = [];
+const emptyTasks: vscode.Task[] = [];
 
 export function enableTaskDetection(): void {
   autoDetectOverride = true;
@@ -95,13 +95,7 @@ export class GradleTaskProvider implements vscode.TaskProvider {
   }
 
   async provideTasks(): Promise<vscode.Task[] | undefined> {
-    try {
-      return await this.provideGradleTasks();
-    } catch (e) {
-      const message = `Error providing gradle tasks: ${e.message}`;
-      console.error(message);
-      this.outputChannel.appendLine(message);
-    }
+    return cachedTasks;
   }
 
   // TODO
@@ -113,49 +107,46 @@ export class GradleTaskProvider implements vscode.TaskProvider {
     return undefined;
   }
 
-  private provideGradleTasks(): Promise<vscode.Task[]> {
-    if (!cachedTasks) {
-      cachedTasks = this.detectGradleTasks();
-    }
-    return cachedTasks;
-  }
-
-  private async detectGradleTasks(): Promise<vscode.Task[]> {
-    const emptyTasks: vscode.Task[] = [];
+  public async refresh(): Promise<vscode.Task[]> {
     const allTasks: vscode.Task[] = [];
     const folders = vscode.workspace.workspaceFolders;
     if (!folders) {
-      return emptyTasks;
-    }
-    try {
-      for (const workspaceFolder of folders) {
-        if (autoDetectOverride || getIsAutoDetectionEnabled(workspaceFolder)) {
-          const projectFolders = await getGradleProjectFolders(workspaceFolder);
-          for (const projectFolder of projectFolders) {
-            allTasks.push(
-              ...(await this.provideGradleTasksForFolder(
-                workspaceFolder,
-                projectFolder
-              ))
+      cachedTasks = emptyTasks;
+    } else {
+      try {
+        for (const workspaceFolder of folders) {
+          if (
+            autoDetectOverride ||
+            getIsAutoDetectionEnabled(workspaceFolder)
+          ) {
+            const projectFolders = await getGradleProjectFolders(
+              workspaceFolder
             );
+            for (const projectFolder of projectFolders) {
+              allTasks.push(
+                ...(await this.provideGradleTasksForFolder(
+                  workspaceFolder,
+                  projectFolder
+                ))
+              );
+            }
           }
         }
+        cachedTasks = allTasks;
+      } catch (e) {
+        cachedTasks = emptyTasks;
+        const message = `Error providing gradle tasks: ${e.message}`;
+        console.error(message);
+        this.outputChannel.appendLine(message);
       }
-      return allTasks;
-    } catch (error) {
-      return Promise.reject(error);
     }
+    return cachedTasks;
   }
 
   private async provideGradleTasksForFolder(
     workspaceFolder: vscode.WorkspaceFolder,
     projectFolder: vscode.Uri
   ): Promise<vscode.Task[]> {
-    const emptyTasks: vscode.Task[] = [];
-    const command = await getGradleWrapperCommandFromPath(projectFolder.fsPath);
-    if (!command) {
-      return emptyTasks;
-    }
     const gradleTasks: GradleTask[] | undefined = await this.getGradleTasks(
       projectFolder
     );
@@ -168,8 +159,7 @@ export class GradleTaskProvider implements vscode.TaskProvider {
         workspaceFolder,
         gradleTask.rootProject,
         vscode.Uri.file(gradleTask.buildFile),
-        projectFolder,
-        ''
+        projectFolder
       )
     );
   }
@@ -180,7 +170,7 @@ export class GradleTaskProvider implements vscode.TaskProvider {
     rootProject: string,
     buildFile: vscode.Uri,
     projectFolder: vscode.Uri,
-    args: string
+    args = ''
   ): vscode.Task {
     const script = gradleTask.path.replace(/^:/, '');
     const definition: GradleTaskDefinition = {
@@ -212,31 +202,12 @@ export class GradleTaskProvider implements vscode.TaskProvider {
 }
 
 export function invalidateTasksCache(): void {
-  cachedTasks = undefined;
+  cachedTasks = emptyTasks;
 }
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function isWorkspaceFolder(value: any): value is vscode.WorkspaceFolder {
   return value && typeof value !== 'number';
-}
-
-export async function getGradleWrapperCommandFromPath(
-  fsPath: string
-): Promise<string> {
-  const platform = process.platform;
-  if (
-    platform === 'win32' &&
-    (await exists(path.join(fsPath, 'gradlew.bat')))
-  ) {
-    return '.\\gradlew.bat';
-  } else if (
-    (platform === 'linux' || platform === 'darwin') &&
-    (await exists(path.join(fsPath, 'gradlew')))
-  ) {
-    return './gradlew';
-  } else {
-    throw new Error('Gradle wrapper executable not found');
-  }
 }
 
 export function getGradleTasksCommand(): string {
@@ -348,10 +319,6 @@ export async function cloneTask(
   client: GradleTasksClient
 ): Promise<vscode.Task | undefined> {
   const folder = task.scope as vscode.WorkspaceFolder;
-  const command = await getGradleWrapperCommandFromPath(folder.uri.fsPath);
-  if (!command) {
-    return undefined;
-  }
   const definition = { ...(task.definition as GradleTaskDefinition), args };
   return createTaskFromDefinition(
     definition as GradleTaskDefinition,
@@ -376,12 +343,4 @@ export async function hasGradleProject(): Promise<boolean> {
     }
   }
   return false;
-}
-
-async function exists(file: string): Promise<boolean> {
-  return new Promise<boolean>(resolve => {
-    fs.exists(file, value => {
-      resolve(value);
-    });
-  });
 }
